Change way to handle with dmg-stats, make teamID int again, use int instead of short

+++ TDS/server/instance/lobby/arena/AllPlayers.cs
@@ -8,7 +8,7 @@ namespace TDS.server.instance.lobby
         private void SetAllPlayersInCountdown ( ) {
             spectatingMe = new Dictionary<Client, List<Client>> ();
             FuncIterateAllPlayers ( ( player, teamID ) => {
-                SetPlayerReadyForRound ( player, (uint) teamID );
+                SetPlayerReadyForRound ( player, teamID );
                 NAPI.ClientEvent.TriggerClientEvent ( player, "onClientCountdownStart" );
                 if ( teamID == 0 )
                     SpectateAllTeams ( player );
